refactor: consolidate chat archive mapping and separate status contracts
