+++ Sources/CrawlBarSelfTest/SelfTest.swift
@@ -23,6 +23,12 @@ enum CrawlBarSelfTest {
         try Self.testNativeConfigRoundTrips()
         try Self.testStatusSecretsLoadFromNativeConfig()
         try Self.testStatusMapperNormalizesCounts()
+        try Self.testStatusMapperGoogleAccountStates()
+        try Self.testStatusMapperWhatsAppStoreStates()
+        try Self.testStatusMapperCrawlKitArchives()
+        try Self.testStatusMapperTelegramFreshness()
+        try Self.testStatusMapperCrawlKitStates()
+        try Self.testStatusMapperGitHubFailures()
         try Self.testStatusMapperTrustsCrawlerState()
         try Self.testStatusMapperNormalizesWacliDoctorOutput()
         try Self.testStatusMapperNormalizesGogAuthStatus()
